Deleting repository and transferring
I'm ending the Didactic World project for now. There was a great idea behind it, but it's not the right time for its release. I will be transferring ownership of Didactic World to ARCHS and its owner, Tim Daniels, who could use some of those anti-raid features of this project for his own organization.